refactor(analysis): Set percentiles depending on the type of analysis request
