+++ lib/actions/analysis/index.js
@@ -10,7 +10,11 @@ import {Map as LeafletMap} from 'leaflet'
 import {createAction} from 'redux-actions'
 import {TextDecoder} from 'text-encoding'
 
-import {ANALYSIS_URL, PROFILE_REQUEST_DEFAULTS} from '../../constants'
+import {
+  ANALYSIS_URL,
+  PROFILE_REQUEST_DEFAULTS,
+  TRAVEL_TIME_PERCENTILES
+} from '../../constants'
 import {
   PERFORMING_ANALYSIS,
   INITIALIZING_CLUSTER,
@@ -93,6 +97,7 @@ export const fetchTravelTimeSurface = () =>
       fromLat: lat,
       fromLon: lon,
       workerVersion,
+      percentiles: TRAVEL_TIME_PERCENTILES,
       projectId: currentProjectId
     }
