+++ lib/actions/analysis/regional.js
@@ -168,7 +168,8 @@ export const createRegionalAnalysis = ({
         method: 'POST',
         body: {
           ...finalProfileRequest,
-          name
+          name,
+          percentiles: [profileRequest.travelTimePercentile || 50]
         }
       },
       url: REGIONAL_URL,
