+++ lib/components/analysis/stacked-percentile.js
@@ -5,7 +5,7 @@ import {scalePow, scaleLinear} from 'd3-scale'
 import {line, area} from 'd3-shape'
 
 import Boxplot from './boxplot'
-import {TRAVEL_TIME_PERCENTILES} from '../../constants/analysis'
+import {TRAVEL_TIME_PERCENTILES} from '../../constants'
 import {PROJECT, BASE, COMPARISON} from './stacked-percentile-selector'
 import messages from '../../utils/messages'
 import MinuteTicks from './minute-ticks'
