+++ lib/components/input.js
@@ -170,6 +170,7 @@ export class Slider extends StateToValue {
           {...props}
           id={id}
           onChange={this._onChange}
+          value={this.state.value}
         />
       </Group>
     )
