--- lib/constants/analysis.js
@@ -1,3 +1,0 @@
-// @flow
-// Percentiles of travel time to request from the backend
-export const TRAVEL_TIME_PERCENTILES = [5, 25, 50, 75, 95]
