+++ lib/constants/index.js
@@ -53,6 +53,12 @@ export const MAP_STATE_SELECT_TO_STOP = 'map state select to stop'
 export const UPDATE_DELAY_MS = 300
 
 /**
+ * Percentiles of travel time to request from the backend. This is for
+ * TRAVEL_TIME_SURFACE requests.
+ */
+export const TRAVEL_TIME_PERCENTILES = [5, 25, 50, 75, 95]
+
+/**
  * Default Profile Request object
  */
 export const PROFILE_REQUEST_DEFAULTS = {
@@ -65,6 +71,7 @@ export const PROFILE_REQUEST_DEFAULTS = {
   maxRides: 4,
   maxTripDurationMinutes: 60,
   monteCarloDraws: 200,
+  percentiles: TRAVEL_TIME_PERCENTILES,
   toTime: 32400,
   transitModes: 'BUS,TRAM,RAIL,SUBWAY',
   travelTimePercentile: 50,
