+++ lib/containers/bookmark-chooser.js
@@ -13,7 +13,7 @@ function mapStateToProps (state, ownProps) {
   return {
     bookmarks: region.bookmarks,
     bookmarkData: selectBookmarkData(state, ownProps),
-    disabled: (!!analysis.isFetchingIsochrone || !analysis.isochroneLonLat),
+    disabled: !!analysis.isFetchingIsochrone,
     selectedBookmark: analysis.selectedBookmark
   }
 }
