+++ lib/selectors/accessibility.js
@@ -1,7 +1,7 @@
 // @flow
 import {createSelector} from 'reselect'
 
-import {TRAVEL_TIME_PERCENTILES} from '../constants/analysis'
+import {TRAVEL_TIME_PERCENTILES} from '../constants'
 import OpportunityDatasets from '../modules/opportunity-datasets'
 
 import type {TravelTimeSurface} from '../types'
