+++ lib/selectors/isochrone.js
@@ -3,7 +3,7 @@ import jsolines from 'jsolines'
 import {Map as LeafletMap} from 'leaflet'
 import {createSelector} from 'reselect'
 
-import {TRAVEL_TIME_PERCENTILES} from '../constants/analysis'
+import {TRAVEL_TIME_PERCENTILES} from '../constants'
 
 const POSITION = TRAVEL_TIME_PERCENTILES.indexOf(50)
