+++ lib/utils/mock-data.js
@@ -12,8 +12,7 @@ import thunkMiddleware from 'redux-thunk'
 import {TextEncoder} from 'text-encoding'
 
 import {responseToSurface} from '../actions/analysis'
-import {ADD_TRIP_PATTERN} from '../constants'
-import {TRAVEL_TIME_PERCENTILES} from '../constants/analysis'
+import {ADD_TRIP_PATTERN, TRAVEL_TIME_PERCENTILES} from '../constants'
 
 export const makeMockStore = configureStore([
   fetchMiddleware,
